move functions "moveToPos" and "MoveBy" into Player Class
