+++ play/src/front/Phaser/Player/Player.ts
@@ -265,16 +265,70 @@ export class Player extends Character {
         return [xDistance / Math.sqrt(distance), yDistance / Math.sqrt(distance)];
     }
 
-    moveBy(x: number, y: number) {
+    /**
+     * Moves the character by the given speed amount.
+     */
+    private moveBy(x: number, y: number) {
+        const body = this.getBody();
+
+        body.setVelocity(x, y);
+
+        if (Math.abs(body.velocity.x) > Math.abs(body.velocity.y)) {
+            if (body.velocity.x < 0) {
+                this._lastDirection = PositionMessage_Direction.LEFT;
+            } else if (body.velocity.x > 0) {
+                this._lastDirection = PositionMessage_Direction.RIGHT;
+            }
+        } else {
+            if (body.velocity.y < 0) {
+                this._lastDirection = PositionMessage_Direction.UP;
+            } else if (body.velocity.y > 0) {
+                this._lastDirection = PositionMessage_Direction.DOWN;
+            }
+        }
         passStatusToOnline();
-        super.moveBy(x, y);
+        this.playAnimation(this._lastDirection, true);
+        this.setDepth(this.y + 16);
+
+        if (this.companion) {
+            this.companion.setTarget(this.x, this.y, this._lastDirection);
+        }
     }
 
-    moveToPos(x: number, y: number) {
+    /**
+     * Moves the character to the given position.
+     */
+    private moveToPos(x: number, y: number) {
+        const oldX = this.x;
+        const oldY = this.y;
+        this.x = x;
+        this.y = y;
+        // The 1.1 ratio to y is applied here because in path finding mode, the player often moves in diagonal.
+        // In diagonal, the amount of x and y are almost equal. This produces a graphical glitch where on one frame,
+        // the player goes left, and on the next frame, the player goes up. This is because the x and y are almost equal.
+        // To fix this, we apply a ratio of 1.1 to y to make sure that the player goes up/down when the y and x are almost equal.
+        if (Math.abs(x - oldX) > Math.abs((y - oldY) * 1.1)) {
+            if (x < oldX) {
+                this._lastDirection = PositionMessage_Direction.LEFT;
+            } else if (x > oldX) {
+                this._lastDirection = PositionMessage_Direction.RIGHT;
+            }
+        } else {
+            if (y < oldY) {
+                this._lastDirection = PositionMessage_Direction.UP;
+            } else if (y > oldY) {
+                this._lastDirection = PositionMessage_Direction.DOWN;
+            }
+        }
         passStatusToOnline();
-        super.moveToPos(x, y);
-    }
+        this.playAnimation(this._lastDirection, true);
 
+        this.setDepth(this.y + 16);
+
+        if (this.companion) {
+            this.companion.setTarget(this.x, this.y, this._lastDirection);
+        }
+    }
     destroy(): void {
         this.unsubscribeVisibilityStore();
         super.destroy();
